Making a request to localhost makes no sense
