+++ lib/explorer.js
@@ -422,7 +422,25 @@ module.exports = {
   {
       var uri = base_url + 'getclaimsfortx?txid=' + tx;
       request({uri: uri, json: true}, function (error, response, body) {
-        return cb(body);
+          var claims = [];
+          var supports = [];
+          body.forEach(function(claim)
+          {
+              var type = ('undefined' == typeof claim['supported claimId']) ? 'claim': 'support';
+              var claimId = (type == 'claim') ? claim['claimId'] : claim['supported claimId'];
+              var c = {claimId: claimId, name: claim.name};
+              if (type == 'claim')
+              {
+                  claims.push(c);
+              }
+
+              else
+              {
+                  supports.push(c);
+              }
+          });
+
+          cb({claims: claims, supports: supports});
       });
   }
